+++ tests/unit_tests/TA_evaluators_tests/test_candle_pattern_TA_evaluator.py
@@ -1,4 +1,4 @@
-from tests.TA_evaluators_tests.abstract_TA_test import AbstractTATest
+from tests.unit_tests.TA_evaluators_tests.abstract_TA_test import AbstractTATest
 from evaluator.TA.momentum_evaluator import CandlePatternMomentumEvaluator
 from config.cst import START_PENDING_EVAL_NOTE
