+++ tests/unit_tests/test_start_bot.py
