+++ tests/unit_tests/trading_tests/test_trader.py
