+++ tests/unit_tests/trading_tests/test_trades_manager.py
